Chore: Migrate lib/rocketchat.js to TS - structure PoC (#3661)
