+++ app/definitions/ITeam.ts
@@ -1,5 +1,5 @@
 // https://github.com/RocketChat/Rocket.Chat/blob/develop/definition/ITeam.ts
-exports.TEAM_TYPE = {
-	PUBLIC: 0,
-	PRIVATE: 1
-};
+export enum TEAM_TYPE {
+	PUBLIC = 0,
+	PRIVATE = 1
+}
